si / cart / main: widen joybus EEPROM type to uint32 and report JDT_EEPROM_NONE when the ROM doesn't use EEPROM
Adapted backport of upstream mupen64plus-core c73a86af 'Fix
"No EEPROM present" Joybus status response'. Their commit
touches different files because of canonical's heavy
src/device/cart/ reorganization vs. our flat per-component
layout, but the on-the-wire fix is identical.

The Joybus EEPROM status response (PIF command 0x00 on the
EEPROM channel) is three bytes: a 24-bit type identifier.
Hardware reports:

    0x008000  -- 4K EEPROM present
    0x00c000  -- 16K EEPROM present
    0xffffff  -- no EEPROM present

Our previous code stored eeprom->id as uint16_t, hard-wired
the high byte (cmd[5] in eeprom_status_command) to 0, and in
main_init unconditionally picked between 0x8000 and 0xc000:

    ROM_SETTINGS.savetype != EEPROM_16KB ? 0x8000 : 0xc000

So a ROM with SaveType=SRAM, FlashRAM, Controller Pak, or
None was reported as having a 4K EEPROM. Most games don't
query the EEPROM unless their save format expects one and
ignored the lie, but homebrew, libdragon-based titles, and
test ROMs that probe joybus do query it and expect the
0xffffff sentinel when there isn't one. They saw 0x008000
and proceeded to read/write a phantom EEPROM, getting
deterministic-but-wrong data back from saved_memory.eeprom
(which the libretro frontend always allocates so the storage
itself is fine -- the problem is the *type advertisement*).

The minimum-change fix is to widen the type field so the
0xffffff sentinel can be carried end-to-end:

  si/eeprom.h
    - struct eeprom::id: uint16_t -> uint32_t
    - init_eeprom() id parameter: uint16_t -> uint32_t
    - Add JDT_EEPROM_NONE / _4K / _16K enum constants so
      callers don't need magic numbers.

  si/eeprom.c
    - init_eeprom() id parameter: uint16_t -> uint32_t
    - eeprom_status_command():
        cmd[5] = (eeprom->id >> 16) & 0xff
      instead of hard-zero. For 4K / 16K this stays 0 (top
      byte of 0x008000 / 0x00c000); for JDT_EEPROM_NONE it
      becomes 0xff, completing the 0xffffff wire pattern.
      Also tighten the existing low/mid byte writes with
      explicit '& 0xff' masks now that the source is wider.

  si/pif.h, si/pif.c
    - init_pif() eeprom_id parameter: uint16_t -> uint32_t.
      Passes straight through to init_eeprom.

  si/si_controller.h, si/si_controller.c
    - init_si() eeprom_id parameter: uint16_t -> uint32_t.
      Passes straight through to init_pif.

  main/device.h, main/device.c
    - init_device() eeprom_id parameter: uint16_t -> uint32_t.
      Passes straight through to init_si.

  main/main.c
    - Add #include for ../si/eeprom.h so JDT_EEPROM_* are in
      scope at the init_device callsite.
    - Compute eeprom_id with an explicit switch on
      ROM_SETTINGS.savetype: EEPROM_4KB -> JDT_EEPROM_4K,
      EEPROM_16KB -> JDT_EEPROM_16K, everything else
      (SRAM / FLASH_RAM / CONTROLLER_PACK / NONE) ->
      JDT_EEPROM_NONE. Pass the resulting eeprom_id into
      init_device.
    - The adjacent eeprom_size argument stays on the
      original (savetype != EEPROM_16KB ? 0x200 : 0x800)
      pattern because the libretro saved_memory.eeprom
      buffer is statically sized and eeprom_size only feeds
      the address-mask path inside eeprom.c, which the
      JDT_EEPROM_NONE case never reaches anyway.

Full MinGW x86_64 cross-build verified clean: every .o
compiles, every cross-TU call into init_eeprom / init_pif /
init_si / init_device links with consistent uint32_t
prototypes. The only link error is the unrelated glsm_ctl
GL-state-manager external satisfied at libretro frontend
link time. Per-TU nm shows the expected T/U pairings on the
widened init functions, no orphan refs.

Behavior change is targeted and small: ROMs that previously
got a 4K EEPROM lie now get the truthful 0xffffff response.
The lie-direction was the bug.

+++ mupen64plus-core/src/main/device.h
@@ -69,7 +69,7 @@ void init_device(
       /* ri */
       uint32_t* dram, size_t dram_size,
       /* si */
-      void* eeprom_user_data, void (*eeprom_save)(void*), uint8_t* eeprom_data, size_t eeprom_size, uint16_t eeprom_id,
+      void* eeprom_user_data, void (*eeprom_save)(void*), uint8_t* eeprom_data, size_t eeprom_size, uint32_t eeprom_id,
       void* af_rtc_user_data, const struct tm* (*af_rtc_get_time)(void*),
       /* sp */
       unsigned int audio_signal,
