+++ mupen64plus-core/src/si/pif.h
@@ -77,7 +77,7 @@ void init_pif(struct pif *pif,
       void (*eeprom_save)(void*),
       uint8_t *eeprom_data,
       size_t eeprom_size,
-      uint16_t eeprom_id,
+      uint32_t eeprom_id,
       void* af_rtc_user_data,
       const struct tm* (*af_rtc_get_time)(void*),
       const uint8_t *ipl3
